backend/requirements.txt
amqp==2.6.1
async-timeout==5.0.1
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.9.0
celery==4.4.7
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.2
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==4.6.11
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.0.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
vine==1.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1
